face_recognition/requirements.txt
fastapi
uvicorn
opencv-python-headless
Pillow
facenet-pytorch
numpy


